+++ frontend/pages/_app.js
@@ -1,0 +1,5 @@
+import '../styles/globals.css';
+
+export default function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
