--- frontend/src/app/globals.css
@@ -1,59 +1,0 @@
-@import "tailwindcss";
-
-:root {
-  --background: #ffffff;
-  --foreground: #000000;
-  --primary: #10b981;
-  --primary-hover: #059669;
-  --border: #e5e7eb;
-  --muted: #f9fafb;
-}
-
-@theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: var(--font-geist-sans);
-  --font-mono: var(--font-geist-mono);
-}
-
-body {
-  background: var(--background);
-  color: var(--foreground);
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-/* Custom button styles */
-.btn-primary {
-  @apply bg-green-600 hover:bg-green-700 text-white font-medium py-2 px-4 rounded-md transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2;
-}
-
-.btn-secondary {
-  @apply bg-gray-100 hover:bg-gray-200 text-gray-900 font-medium py-2 px-4 rounded-md transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2;
-}
-
-/* Custom input styles */
-.input-field {
-  @apply w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-green-500 transition-colors duration-200;
-}
-
-/* Custom card styles */
-.card {
-  @apply bg-white rounded-lg shadow-sm border border-gray-200;
-}
-
-/* Status badge styles */
-.badge-success {
-  @apply bg-green-100 text-green-800 px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
-
-.badge-pending {
-  @apply bg-yellow-100 text-yellow-800 px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
-
-.badge-processing {
-  @apply bg-blue-100 text-blue-800 px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
-
-.badge-error {
-  @apply bg-red-100 text-red-800 px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
